+++ threedi_schematisation_editor/custom_widgets/__init__.py
@@ -11,7 +11,13 @@ from qgis.PyQt.QtGui import QStandardItem, QStandardItemModel
 from qgis.PyQt.QtWidgets import QComboBox, QInputDialog, QTableWidgetItem
 
 import threedi_schematisation_editor.data_models as dm
-from threedi_schematisation_editor.custom_tools import ColumnImportMethod, CulvertImportConfig, CulvertsImporter
+from threedi_schematisation_editor.custom_tools import (
+    ColumnImportMethod,
+    CulvertsImporter,
+    OrificesImporter,
+    StructuresImportConfig,
+    WeirsImporter,
+)
 from threedi_schematisation_editor.utils import (
     NULL_STR,
     QUOTED_NULL,
@@ -24,7 +30,7 @@ from threedi_schematisation_editor.utils import (
 )
 
 ps_basecls, ps_uicls = uic.loadUiType(os.path.join(os.path.dirname(__file__), "ui", "projection_selection.ui"))
-ic_basecls, ic_uicls = uic.loadUiType(os.path.join(os.path.dirname(__file__), "ui", "import_culverts.ui"))
+ic_basecls, ic_uicls = uic.loadUiType(os.path.join(os.path.dirname(__file__), "ui", "import_structures.ui"))
 vm_basecls, vm_uicls = uic.loadUiType(os.path.join(os.path.dirname(__file__), "ui", "attribute_value_map.ui"))
 
 
@@ -118,18 +124,12 @@ class AttributeValueMapDialog(vm_basecls, vm_uicls):
             skipped_rows = 0
             for i, source_value in enumerate(sorted(unique_values), start=last_row_number):
                 source_value_str = self.format_value_map_data(source_value)
-                if isinstance(source_value, str):
-                    source_value = source_value_str
-                elif source_value == NULL:
-                    source_value = source_value_str
-                else:
-                    source_value = source_value_str
-                if source_value in existing_values:
+                if source_value_str in existing_values:
                     skipped_rows += 1
                     continue
                 new_row_number = i - skipped_rows
                 self.value_map_table.insertRow(new_row_number)
-                self.value_map_table.setItem(new_row_number, self.SRC_COLUMN_IDX, QTableWidgetItem(source_value))
+                self.value_map_table.setItem(new_row_number, self.SRC_COLUMN_IDX, QTableWidgetItem(source_value_str))
                 self.value_map_table.setItem(new_row_number, self.DST_COLUMN_IDX, QTableWidgetItem(QUOTED_NULL))
 
     @staticmethod
@@ -155,36 +155,52 @@ class AttributeValueMapDialog(vm_basecls, vm_uicls):
         self.update_value_map_button(self.pressed_button, new_value_map)
 
 
-class ImportCulvertsDialog(ic_basecls, ic_uicls):
-    """Dialog for the importing culverts tool."""
+class ImportStructuresDialog(ic_basecls, ic_uicls):
+    """Dialog for the importing structures tool."""
 
-    def __init__(self, model_gpkg, layer_manager, uc, parent=None):
+    STRUCTURE_IMPORTERS = {
+        dm.Culvert: CulvertsImporter,
+        dm.Orifice: OrificesImporter,
+        dm.Weir: WeirsImporter,
+    }
+
+    def __init__(self, structure_model_cls, model_gpkg, layer_manager, uc, parent=None):
         super().__init__(parent)
         self.setupUi(self)
+        self.structure_model_cls = structure_model_cls
         self.model_gpkg = model_gpkg
         self.layer_manager = layer_manager
         self.uc = uc
-        self.import_configuration = CulvertImportConfig()
-        self.culvert_model = QStandardItemModel()
-        self.culvert_tv.setModel(self.culvert_model)
+        self.import_configuration = StructuresImportConfig(self.structure_model_cls)
+        self.structure_importer_cls = self.STRUCTURE_IMPORTERS[structure_model_cls]
+        self.structure_model = QStandardItemModel()
+        self.structure_tv.setModel(self.structure_model)
         self.connection_node_model = QStandardItemModel()
         self.connection_node_tv.setModel(self.connection_node_model)
         self.data_models_tree_views = {
-            dm.Culvert: (self.culvert_tv, self.culvert_model),
+            self.structure_model_cls: (self.structure_tv, self.structure_model),
             dm.ConnectionNode: (self.connection_node_tv, self.connection_node_model),
         }
-        self.culvert_layer_cbo.setFilters(QgsMapLayerProxyModel.LineLayer)
+        self.structure_layer_cbo.setFilters(QgsMapLayerProxyModel.LineLayer)
         self.populate_conversion_settings_widgets()
-        self.culvert_layer_cbo.layerChanged.connect(self.on_layer_changed)
+        self.structure_layer_cbo.layerChanged.connect(self.on_layer_changed)
         self.create_nodes_cb.stateChanged.connect(self.on_create_nodes_change)
         self.save_pb.clicked.connect(self.save_import_settings)
         self.load_pb.clicked.connect(self.load_import_settings)
-        self.run_pb.clicked.connect(self.run_import_culverts)
+        self.run_pb.clicked.connect(self.run_import_structures)
         self.close_pb.clicked.connect(self.close)
+        self.setup_labels()
+
+    def setup_labels(self):
+        structure_name = self.structure_model_cls.__layername__
+        structure_name_lower = structure_name.lower()
+        self.setWindowTitle(self.windowTitle().format(structure_name_lower))
+        self.structure_layer_label.setText(self.structure_layer_label.text().format(structure_name_lower))
+        self.tab_widget.setTabText(0, self.tab_widget.tabText(0).format(structure_name))
 
     @property
     def source_layer(self):
-        return self.culvert_layer_cbo.currentLayer()
+        return self.structure_layer_cbo.currentLayer()
 
     def on_create_nodes_change(self, is_checked):
         self.connection_node_tab.setEnabled(is_checked)
@@ -194,7 +210,7 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
         if layer:
             layer_field_names += [field.name() for field in layer.fields()]
         source_attribute_widgets = self.get_column_widgets(
-            CulvertImportConfig.SOURCE_ATTRIBUTE_COLUMN_IDX, dm.Culvert, dm.ConnectionNode
+            StructuresImportConfig.SOURCE_ATTRIBUTE_COLUMN_IDX, self.structure_model_cls, dm.ConnectionNode
         )
         for combobox in chain.from_iterable(source_attribute_widgets.values()):
             combobox.clear()
@@ -244,7 +260,7 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
         return column_widgets
 
     def populate_conversion_settings_widgets(self):
-        widgets_to_add = self.import_configuration.culvert_widgets()
+        widgets_to_add = self.import_configuration.structure_widgets()
         for model_cls, (tree_view, tree_view_model) in self.data_models_tree_views.items():
             tree_view_model.clear()
             tree_view_model.setHorizontalHeaderLabels(self.import_configuration.config_header)
@@ -258,16 +274,18 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
         self.on_layer_changed(self.source_layer)
 
     def connect_configuration_widgets(self):
-        for model_cls in [dm.Culvert, dm.ConnectionNode]:
+        for model_cls in [self.structure_model_cls, dm.ConnectionNode]:
             tree_view, tree_view_model = self.data_models_tree_views[model_cls]
             for row_idx, field_name in enumerate(model_cls.__annotations__.keys()):
-                method_item = tree_view_model.item(row_idx, CulvertImportConfig.METHOD_COLUMN_IDX)
+                method_item = tree_view_model.item(row_idx, StructuresImportConfig.METHOD_COLUMN_IDX)
                 method_index = method_item.index()
                 method_combobox = tree_view.indexWidget(method_index)
-                source_attribute_item = tree_view_model.item(row_idx, CulvertImportConfig.SOURCE_ATTRIBUTE_COLUMN_IDX)
+                source_attribute_item = tree_view_model.item(
+                    row_idx, StructuresImportConfig.SOURCE_ATTRIBUTE_COLUMN_IDX
+                )
                 source_attribute_index = source_attribute_item.index()
                 source_attribute_combobox = tree_view.indexWidget(source_attribute_index)
-                value_map_item = tree_view_model.item(row_idx, CulvertImportConfig.VALUE_MAP_COLUMN_IDX)
+                value_map_item = tree_view_model.item(row_idx, StructuresImportConfig.VALUE_MAP_COLUMN_IDX)
                 value_map_index = value_map_item.index()
                 value_map_button = tree_view.indexWidget(value_map_index)
                 method_combobox.currentTextChanged.connect(
@@ -283,13 +301,13 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
 
     def collect_settings(self):
         import_settings = {
-            "target_layer": dm.Culvert.__tablename__,
+            "target_layer": self.structure_model_cls.__tablename__,
             "conversion_settings": {
                 "use_snapping": self.snap_gb.isChecked(),
                 "snapping_distance": self.snap_dsb.value(),
                 "create_connection_nodes": self.create_nodes_cb.isChecked(),
             },
-            "fields": self.collect_fields_settings(dm.Culvert),
+            "fields": self.collect_fields_settings(self.structure_model_cls),
             "connection_node_fields": self.collect_fields_settings(dm.ConnectionNode),
         }
         return import_settings
@@ -307,10 +325,10 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
                 if isinstance(widget, QComboBox):
                     key_value = (
                         widget.currentText()
-                        if column_idx == CulvertImportConfig.SOURCE_ATTRIBUTE_COLUMN_IDX
+                        if column_idx == StructuresImportConfig.SOURCE_ATTRIBUTE_COLUMN_IDX
                         else widget.currentData()
                     )
-                elif column_idx == CulvertImportConfig.VALUE_MAP_COLUMN_IDX:
+                elif column_idx == StructuresImportConfig.VALUE_MAP_COLUMN_IDX:
                     key_value = widget.value_map
                     if not key_value:
                         continue
@@ -318,7 +336,7 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
                     key_value = widget.text()
                     if not key_value:
                         continue
-                    if column_idx == CulvertImportConfig.DEFAULT_VALUE_COLUMN_IDX and key_name != NULL_STR:
+                    if column_idx == StructuresImportConfig.DEFAULT_VALUE_COLUMN_IDX and key_name != NULL_STR:
                         key_value = field_type(key_value)
                 single_field_config[key_name] = key_value
             fields_settings[field_name] = single_field_config
@@ -344,7 +362,7 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
                     else:
                         if key_name == "method":
                             widget.setCurrentText(enum_entry_name_format(ColumnImportMethod(key_value).name))
-                        elif key_name == "default_value":
+                        elif key_name == "default_value" and field_type != bool:
                             widget.setCurrentText(enum_entry_name_format(field_type(key_value).name))
                         else:
                             widget.setCurrentText(str(key_value))
@@ -379,7 +397,7 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
             self.snap_gb.setChecked(conversion_settings.get("use_snapping", False))
             self.snap_dsb.setValue(conversion_settings.get("snapping_distance", 0.1))
             self.create_nodes_cb.setChecked(conversion_settings.get("create_connection_nodes", False))
-            self.update_fields_settings(dm.Culvert, import_settings["fields"])
+            self.update_fields_settings(self.structure_model_cls, import_settings["fields"])
             try:
                 connection_node_fields = import_settings["connection_node_fields"]
                 self.update_fields_settings(dm.ConnectionNode, connection_node_fields)
@@ -390,13 +408,13 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
             self.uc.show_error(f"Import failed due to the following error:\n{e}", self)
 
     def missing_source_fields(self):
-        data_models = [dm.Culvert]
+        data_models = [self.structure_model_cls]
         if self.create_nodes_cb.isChecked():
             data_models.append(dm.ConnectionNode)
-        field_labels = self.get_column_widgets(CulvertImportConfig.FIELD_NAME_COLUMN_IDX, *data_models)
-        method_widgets = self.get_column_widgets(CulvertImportConfig.METHOD_COLUMN_IDX, *data_models)
+        field_labels = self.get_column_widgets(StructuresImportConfig.FIELD_NAME_COLUMN_IDX, *data_models)
+        method_widgets = self.get_column_widgets(StructuresImportConfig.METHOD_COLUMN_IDX, *data_models)
         source_attribute_widgets = self.get_column_widgets(
-            CulvertImportConfig.SOURCE_ATTRIBUTE_COLUMN_IDX, *data_models
+            StructuresImportConfig.SOURCE_ATTRIBUTE_COLUMN_IDX, *data_models
         )
         missing_fields = {}
         for model_cls in data_models:
@@ -418,7 +436,7 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
                 missing_fields[model_cls] = model_missing_fields
         return missing_fields
 
-    def run_import_culverts(self):
+    def run_import_structures(self):
         missing_fields = self.missing_source_fields()
         if missing_fields:
             missing_fields_lines = []
@@ -432,30 +450,30 @@ class ImportCulvertsDialog(ic_basecls, ic_uicls):
             )
             return
         source_layer = self.source_layer
-        culvert_handler = self.layer_manager.model_handlers[dm.Culvert]
+        structures_handler = self.layer_manager.model_handlers[self.structure_model_cls]
         node_handler = self.layer_manager.model_handlers[dm.ConnectionNode]
-        culvert_layer = culvert_handler.layer
+        structure_layer = structures_handler.layer
         node_layer = node_handler.layer
         selected_feat_ids = None
         if self.selected_only_cb.isChecked():
             selected_feat_ids = source_layer.selectedFeatureIds()
         import_settings = self.collect_settings()
         try:
-            culvert_handler.disconnect_handler_signals()
+            structures_handler.disconnect_handler_signals()
             node_handler.disconnect_handler_signals()
-            culverts_importer = CulvertsImporter(
-                source_layer, self.model_gpkg, import_settings, culvert_layer=culvert_layer, node_layer=node_layer
+            structures_importer = self.structure_importer_cls(
+                source_layer, self.model_gpkg, import_settings, structure_layer=structure_layer, node_layer=node_layer
             )
-            success, commit_errors = culverts_importer.import_culverts(selected_ids=selected_feat_ids)
+            success, commit_errors = structures_importer.import_structures(selected_ids=selected_feat_ids)
             if not success:
                 commit_errors_message = "\n".join(commit_errors)
                 self.uc.show_warn(f"Import failed due to the following errors:\n{commit_errors_message}")
             else:
-                self.uc.show_info(f"Culverts successfully imported.", self)
+                self.uc.show_info(f"Features successfully imported.", self)
         except Exception as e:
             self.uc.show_error(f"Import failed due to the following error:\n{e}", self)
         finally:
-            culvert_handler.connect_handler_signals()
+            structures_handler.connect_handler_signals()
             node_handler.connect_handler_signals()
-        for layer in [culvert_layer, node_layer]:
+        for layer in [structure_layer, node_layer]:
             layer.triggerRepaint()