+++ threedi_schematisation_editor/processing/__init__.py
@@ -5,7 +5,7 @@ from qgis.core import QgsProcessingProvider
 from qgis.PyQt.QtGui import QIcon
 
 from threedi_schematisation_editor.processing.algorithms_1d2d import GenerateExchangeLines
-from threedi_schematisation_editor.processing.algorithms_conversion import ImportCulverts
+from threedi_schematisation_editor.processing.algorithms_conversion import ImportCulverts, ImportOrifices, ImportWeirs
 
 
 class ThreediSchematisationEditorProcessingProvider(QgsProcessingProvider):
@@ -33,6 +33,6 @@ class ThreediSchematisationEditorProcessingProvider(QgsProcessingProvider):
         self.algorithms_list = None
 
     def loadAlgorithms(self):
-        self.algorithms_list = [GenerateExchangeLines(), ImportCulverts()]
+        self.algorithms_list = [GenerateExchangeLines(), ImportCulverts(), ImportOrifices(), ImportWeirs()]
         for alg in self.algorithms_list:
             self.addAlgorithm(alg)